Move some things to the service
